+++ know-your-votes/CHANGELOG.md
@@ -14,7 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Updated Main Home Section with responsiveness
-- Update css for svg map 
+- Add responsive democrat and republican css for states
 
 ### Fixed